DTNN-609 Reorganized unit test files to follow standard organization
