+++ unit-test/bpnode_cmds_tests.c
@@ -17,7 +17,7 @@
  ************************************************************************/
 
 /*
-** File: coveragetest_bpnode_cmds.c
+** File: bpnode_cmds_tests.c
 **
 ** Purpose:
 ** Coverage Unit Test cases for the BPNode Application
@@ -37,7 +37,7 @@
  */
 
 #include "bplib.h"
-#include "bpnode_coveragetest_common.h"
+#include "bpnode_test_utils.h"
 #include "bpnode_app.h"
 #include "bpnode_dispatch.h"
 #include "bpnode_cmds.h"
