--- unit-test/common/setup.c
@@ -1,55 +1,0 @@
-/************************************************************************
- * NASA Docket No. GSC-18,719-1, and identified as “core Flight System: Bootes”
- *
- * Copyright (c) 2020 United States Government as represented by the
- * Administrator of the National Aeronautics and Space Administration.
- * All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License. You may obtain
- * a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- ************************************************************************/
-
-/*
-** Purpose:
-** Coverage Unit Test cases for the BPNode Application
-**
-** Notes:
-** This implements various test cases to exercise all code
-** paths through all functions defined in the BPNode application.
-**
-** It is primarily focused at providing examples of the various
-** stub configurations, hook functions, and wrapper calls that
-** are often needed when coercing certain code paths through
-** complex functions.
-*/
-
-/*
- * Includes
- */
-#include "common_types.h"
-
-#include "setup.h"
-
-#include "utassert.h"
-#include "uttest.h"
-#include "utstubs.h"
-
-/*
- * Setup function prior to every test
- */
-void BPNode_UT_Setup(void)
-{
-    UT_ResetState(0);
-}
-
-/*
- * Teardown function after every test
- */
-void BPNode_UT_TearDown(void) {}
